perf(kws): 借鉴白龙马 - Float32 直发 + 1600 样本块 + 关浏览器音频处理
前端:
- 新增 /audio/pcm-worklet-kws.js: KWS 专用 Float32 worklet (1600 样本/块 = 100ms@16k)
- pcm-worklet.ts: 导出 KWS Float32 worklet 源码 (Blob URL 加载)
- kwsMonitor.ts: 切换到 pcm-worklet-kws (Float32),发送 Float32Array
  - 关掉 noiseSuppression (之前 true)
  - 关掉 autoGainControl (之前未设,默认 true)
  - 借鉴白龙马 wake-probe.html: 'KWS 要原始音频,AEC/NS/AGC 全关'
- kwsClient.ts: sendAudio(pcm: Float32Array),不再做 Int16 转换

后端 (KwsService.java):
- 输入格式 Int16 LE → Float32 LE (4 字节/样本)
- 移除 sessionSampleBuffer / sessionBufferFill 累积逻辑
  - 之前要攒到 8000 样本 (500ms) 才喂,有 500ms 延迟
  - 现在每块 1600 样本直发 acceptWaveform (100ms 延迟)
- pcm16kMonoInt16ToFloat → pcm16kMonoFloat32BytesToFloat (ByteBuffer 包装)
- 简化 acceptFrame: 字节→float[]→acceptWaveform→while isReady decode

测试 (KwsServiceTest.java):
- WAV 还是 Int16,新增 int16LeToFloat32Le() 转换
- frame size 3200 字节 (Int16 1600 样本) → 6400 字节 (Float32 1600 样本)

不动: STT 的 recorder.ts / pcm-worklet.js (Int16) 保持原样

借鉴来源: docs/white-dragon-horse-kws-analysis.md §6.1

+++ frontend/src/api/kwsClient.ts
@@ -1,6 +1,10 @@
 // kwsClient.ts -- KWS 唤醒词专用 WebSocket 客户端 (v3 新增)
 // 跟现有 VoiceClient 平行,连 /ws/kws 而不是 /ws/audio
 // 完全独立,不修改任何现有文件
+//
+// 改动 (2026-07-25): PCM 格式从 Int16 改成 Float32 (直接发,不做 Int16 转换)
+//   理由: KWS 模型在 float 上训练,Float32→Int16 会引入量化噪声
+//   借鉴白龙马 wake-probe.html + voice-core.js 的 Float32 直发模式
 
 type Handler = (...args: any[]) => void
 
@@ -34,8 +38,9 @@ export class KwsClient {
     }
   }
 
-  /** 发 PCM 二进制帧 (16kHz mono int16 LE) */
-  sendAudio(pcm: Int16Array): void {
+  /** 发 Float32 PCM 二进制帧 (16kHz mono float32 LE,4 字节/样本)。
+   *  借鉴白龙马:KWS 模型在 float 上训练,不做 Int16 转换。 */
+  sendAudio(pcm: Float32Array): void {
     if (this.ws?.readyState === WebSocket.OPEN) {
       this.ws.send(pcm.buffer)
     }
